+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@toss/utils": "^1.6.1",
     "@types/react-slick": "^0.23.13",
     "@vercel/analytics": "^1.5.0",
     "axios": "^1.7.9",
@@ -16,6 +17,7 @@
     "crypto-js": "^4.2.0",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
+    "fitty": "^2.4.2",
     "framer-motion": "^12.4.3",
     "gsap": "^3.12.7",
     "jsonwebtoken": "^9.0.2",
