+++ src/app/(route)/home/_components/StatusSection/StatCounter.tsx
