+++ src/app/(route)/home/_types/cardData.ts
@@ -1,0 +1,11 @@
+// 카드 데이터 타입 정의
+export interface CardData {
+  id: number;
+  title: string;
+  image: string;
+  tag: string;
+  descriptions: string[];
+  link: string;
+  bgColor1: string;
+  bgColor2: string;
+}
