+++ src/app/(route)/home/page.tsx
@@ -1,14 +1,14 @@
 "use client";
 
-import HeroSection from "@/src/app/(route)/home/_components/HeroSection";
-import IntroductionSection from "@/src/app/(route)/home/_components/IntroductionSection";
-import StatusSection from "@/src/app/(route)/home/_components/StatusSection";
-import CTASection from "@/src/app/(route)/home/_components/CTASection";
+import HeroTitleSection from "./_components/HeroTitleSection";
+import IntroductionSection from "./_components/IntroductionSection/indext";
+import StatusSection from "./_components/StatusSection";
+import CTASection from "./_components/CTASection";
 
 export default function Home() {
   return (
     <div className="min-h-screen text-gray-800 overflow-hidden">
-      <HeroSection />
+      <HeroTitleSection />
       <IntroductionSection />
       <StatusSection />
       <CTASection />
