+++ src/shared/components/index.ts
@@ -1,0 +1,2 @@
+export { default as AnimatedTitle } from "./AnimatedTitle";
+export { default as AnimatedTitleDescription } from "./AnimatedTitleDescription";
