+++ src/shared/hook/useGsapAnimaions.ts
@@ -1,0 +1,36 @@
+import { useEffect } from "react";
+import gsap from "gsap";
+import { ScrollTrigger } from "gsap/ScrollTrigger";
+
+/**
+ * 애니메이션 초기화 훅
+ * @returns 애니메이션 초기화 훅
+ */
+export default function useGsapAnimations() {
+  useEffect(() => {
+    gsap.registerPlugin(ScrollTrigger);
+
+    const animations = gsap.context(() => {
+      gsap.utils.toArray(".gsap-fade-in").forEach((element: any, index) => {
+        gsap.fromTo(
+          element,
+          { opacity: 0, y: 80 },
+          {
+            opacity: 1,
+            y: 0,
+            duration: 0.9,
+            delay: index * 0.1,
+            ease: "power3.out",
+            scrollTrigger: {
+              trigger: element,
+              start: "top 90%",
+              toggleActions: "play none none reverse",
+            },
+          }
+        );
+      });
+    });
+
+    return () => animations.revert();
+  }, []);
+}
